pyproject.toml
[tool.poetry]
name = "KaTrain"
version = "0.0.0"  # Placeholder. The poetry-dynamic-versioning sets the version
description = "Go/Baduk/Weiqi playing and teaching app with a variety of AIs"
authors = ["Sander Land"]
license = "MIT"
homepage = "https://github.com/sanderland/katrain"
readme = "README.md"
packages = [{include = "katrain"}]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Games/Entertainment :: Board Games"
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
kivy = {extras = ["full"], version = ">=2.1.0"}
kivymd = "==0.104.1"  # TODO: upgrade this, the latest version is 1.1.1
ffpyplayer = "*"
urllib3 = "*"
pygame = {version = "^2.0", markers = "platform_system == 'Darwin'"}    # some mac versions need this for kivy
screeninfo = {version = "^0.8.1", markers = "platform_system != 'Darwin'"}  # for screen resolution, has problems on macos
chardet = "^5.2.0"  # for automatic encoding detection
# Avoid PyPI/Poetry problem: https://github.com/python-poetry/poetry/issues/9293
docutils = ">=0.21.2"
openai = "^1.58.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
polib = "^1.2.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"

[tool.poetry.scripts]
katrain = "katrain.__main__:run_app"

[tool.black]
line-length = 120

[tool.poetry-dynamic-versioning]
enable = true

[tool.poetry-dynamic-versioning.from-file]
source = "katrain/core/constants.py"
pattern = "^VERSION\\s*=\\s*\"(.*)\""

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
